requirements.txt
asgiref==3.7.2
Django==4.2.4
mysqlclient==2.2.0
mysql-connector-python==8.0.23
sqlparse==0.4.4
tzdata==2023.3
